Separate CodingQuestion into another component.
